+++ bw2data/filesystem.py
@@ -2,28 +2,8 @@
 import hashlib
 import os
 import re
-import unicodedata
 
 re_slugify = re.compile(r"[^\w\s-]", re.UNICODE)
-
-
-def safe_filename(string, add_hash=True):
-    """Convert arbitrary strings to make them safe for filenames. Substitutes strange characters, and uses unicode normalization.
-
-    if `add_hash`, appends hash of `string` to avoid name collisions.
-
-    From http://stackoverflow.com/questions/295135/turn-a-string-into-a-valid-filename-in-python"""
-    safe = re.sub(
-        r"[-\s]+",
-        "-",
-        str(re_slugify.sub("", unicodedata.normalize("NFKD", str(string))).strip()),
-    )
-    if add_hash:
-        if isinstance(string, str):
-            string = string.encode("utf8")
-        return safe + "." + hashlib.md5(string).hexdigest()
-    else:
-        return safe
 
 
 def create_dir(dirpath):